fix: Bucket not empty error on rollback (#83)
Attempt to fix errors where a bucket is being deleted on rollback, but it's nto empty. The auto object remover should delete everything before the bucket is deleted. I think what's happening is that some rollback of a packager writes to the bucket after the auto-deleter gets rolled-back. This attempts to fix the issue by creating a dependency on the deleter from the packager.